Refactor: decouple adding/creating tasks
